# openapi.json
{
  "components": {
    "schemas": {
      "CommitRequest": {
        "additionalProperties": false,
        "properties": {
          "added_back": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Added Back",
            "type": "array"
          },
          "injection_id": {
            "format": "uuid",
            "title": "Injection Id",
            "type": "string"
          },
          "removed": {
            "items": {
              "$ref": "#/components/schemas/RemovedMemory"
            },
            "title": "Removed",
            "type": "array"
          }
        },
        "required": [
          "injection_id",
          "removed",
          "added_back"
        ],
        "title": "CommitRequest",
        "type": "object"
      },
...
      "FeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "injection_id": {
            "format": "uuid",
            "title": "Injection Id",
            "type": "string"
          },
          "memory_id": {
            "format": "uuid",
            "title": "Memory Id",
            "type": "string"
          },
          "signal": {
            "enum": [
              "mid_thread_removed",
              "cited"
            ],
            "title": "Signal",
            "type": "string"
          }
        },
        "required": [
          "injection_id",
          "memory_id",
          "signal"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
...
      "PrepareRequest": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "agent_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Kind"
          },
          "machine_id": {
            "title": "Machine Id",
            "type": "string"
          },
          "model_context_tokens": {
            "title": "Model Context Tokens",
            "type": "integer"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "string"
          },
          "project_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Key"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "thread_id": {
            "format": "uuid",
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "agent_id",
          "machine_id",
          "principal_id",
          "prompt",
          "model_context_tokens"
        ],
        "title": "PrepareRequest",
        "type": "object"
      },
      "RemovedMemory": {
        "additionalProperties": false,
        "properties": {
          "memory_id": {
            "format": "uuid",
            "title": "Memory Id",
            "type": "string"
          },
          "reason": {
            "enum": [
              "not_relevant",
              "wrong",
              "never"
            ],
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "memory_id",
          "reason"
        ],
        "title": "RemovedMemory",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "StaticBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "N8 Spine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/feedback": {
      "post": {
        "operationId": "feedback_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Bearer token missing or invalid"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Request does not match the endpoint contract"
          },
          "501": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Agent Zero contract stub"
          }
        },
        "security": [
          {
            "StaticBearer": []
          }
        ],
        "summary": "Feedback",
        "tags": [
          "injection"
        ]
      }
    },
    "/v1/inject/commit": {
      "post": {
        "operationId": "commit_v1_inject_commit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Bearer token missing or invalid"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Request does not match the endpoint contract"
          },
          "501": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Agent Zero contract stub"
          }
        },
        "security": [
          {
            "StaticBearer": []
          }
        ],
        "summary": "Commit",
        "tags": [
          "injection"
        ]
      }
    },
    "/v1/inject/prepare": {
      "post": {
        "operationId": "prepare_v1_inject_prepare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Bearer token missing or invalid"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Request does not match the endpoint contract"
          },
          "501": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Agent Zero contract stub"
          }
        },
        "security": [
          {
            "StaticBearer": []
          }
        ],
        "summary": "Prepare",
        "tags": [
          "injection"
        ]
      }
    },
...
  }
}